python-scripts/requirements.txt
setuptools>=41.0.0
pip>=19.0.0
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.4
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
progressbar2==3.46.1
entrypoints==0.3
gast==0.3.3
google-pasta==0.1.8
grpcio==1.29.0
h5py==2.10.0
idna==2.8
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
jedi==0.14.0
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.0
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.17.3
#opencv-python==4.1.2.30
opencv-python
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.3
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
# pyzmq==18.0.2
pyzmq
qtconsole==4.5.1
requests==2.22.0
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
sympy==1.6
# tensorflow==2.2.0
tensorflow
# tensorboard==2.2.2
# tensorboard
# tensorflow-estimator==2.2.0
# tensorflow-estimator
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.0
wrapt==1.11.2
Flask==1.1.2
gevent
